requirements-test.txt
mysqlclient>=1.3,<1.4
psycopg2>=2.7,<2.8
psycopg2-binary>=2.7,<2.8
django-debug-toolbar==1.9.1
python-dateutil>2.0,<3.0
selenium>=3.0,<4.0
